Revert "begin introducing receipts"
This reverts commit e6301e5ed1283f3ed5ce0ad26ea10b487678caf9.

This was cute because it gave Vendor a Functor instance

It was intended to implement 'match'; the receipt would be
a running total of the number of characters committed to
a proxy vendor that doesn't actually commit upstream.

I think this mechanism is superfluous because the same thing
can be achieved by adding an additional type of request to
the vendor's downstream interface to achieve the same
extraction of state that receipts were intended to provide.
Moreover, going through the interface can be more efficient,
as the "receipt" can then be given only when it is needed,
not upon every order whether it is wanted or not.